% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medida_pulgadas.R
\name{pulgadas_a_centimetros}
\alias{pulgadas_a_centimetros}
\title{Convierte pulgadas a centímetros}
\usage{
pulgadas_a_centimetros(medida_pulgadas)
}
\arguments{
\item{medida_pulgadas}{Un valor numérico que representa la medida en pulgadas.}
}
\value{
Un valor numérico correspondiente a la medida en centímetros.
}
\description{
Esta función toma una medida en pulgadas y la convierte a centímetros.
Si el valor ingresado no es numérico, genera un error utilizando
\code{cli::cli_abort()} indicando el tipo de dato ingresado.
}
\examples{
pulgadas_a_centimetros(10)
# [1] 25.4

}
