% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x = 0, y = 0)
}
\arguments{
\item{x}{un valor vector o numerico}

\item{y}{un valor vector o numerico}
}
\value{
Un valor o vector numerico con la suma de x mas y
}
\description{
Suma de dos numeros que deben ser positivos.
}
\examples{
suma(10,1)
suma(3,2)

}
